Move global.cpp to native/gfx/d3d9_shader.cpp
