+++ GPU/Directx9/ShaderManagerDX9.cpp
@@ -20,7 +20,7 @@
 #endif
 
 #include <map>
-#include "helper/global.h"
+#include "gfx/d3d9_shader.h"
 #include "base/logging.h"
 #include "i18n/i18n.h"
 #include "math/lin/matrix4x4.h"
