+++ GPU/GPU.vcxproj.filters
@@ -105,9 +105,6 @@
     <ClInclude Include="Directx9\helper\dx_state.h">
       <Filter>DirectX9\helper</Filter>
     </ClInclude>
-    <ClInclude Include="Directx9\helper\global.h">
-      <Filter>DirectX9\helper</Filter>
-    </ClInclude>
     <ClInclude Include="Common\TextureDecoder.h">
       <Filter>Common</Filter>
     </ClInclude>
@@ -312,9 +309,6 @@
       <Filter>Common</Filter>
     </ClCompile>
     <ClCompile Include="Directx9\helper\dx_state.cpp">
-      <Filter>DirectX9\helper</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\helper\global.cpp">
       <Filter>DirectX9\helper</Filter>
     </ClCompile>
     <ClCompile Include="Common\TextureDecoder.cpp">
